+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 📦 **settings**: Added settings app
 - 🔧 **scripts**: enhance commit helper with CLI support
 - 🧪 add simple test file
+- 🔧 **Settings**: Reworked settings
 
 ## [0.2.0] - 2025-08-07